# tests/resources/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_put_blob": {
        "properties": {
          "blob": {
            "format": "binary",
            "title": "Blob",
            "type": "string"
          }
        },
        "required": [
          "blob"
        ],
        "title": "Body_put_blob",
        "type": "object"
      },
      "ChangePasswordRequest": {
        "description": "Change a password",
        "properties": {
          "new1": {
            "title": "New1",
            "type": "string"
          },
          "old": {
            "title": "Old",
            "type": "string"
          }
        },
        "required": [
          "old",
          "new1"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
...
      "Entry": {
        "description": "Directory entry",
        "properties": {
          "attrs": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Attrs",
            "type": "object"
          },
          "changed": {
            "items": {
              "type": "string"
            },
            "title": "Changed",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "attrs",
          "meta"
        ],
        "title": "Entry",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBasic": {
        "scheme": "basic",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "LDAP UI",
    "version": "0.10.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/blob/{attr}/{index}/{dn}": {
      "delete": {
        "description": "Remove a binary attribute",
        "operationId": "delete_blob",
        "parameters": [
          {
            "in": "path",
            "name": "attr",
            "required": true,
            "schema": {
              "title": "Attr",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Delete Blob",
        "tags": [
          "Editing"
        ]
      },
      "put": {
        "description": "Upload a binary attribute",
        "operationId": "put_blob",
        "parameters": [
          {
            "in": "path",
            "name": "attr",
            "required": true,
            "schema": {
              "title": "Attr",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_put_blob"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Put Blob",
        "tags": [
          "Editing"
        ]
      }
    },
    "/api/change-password/{dn}": {
      "post": {
        "description": "Update passwords",
        "operationId": "post_change_password",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "Editing"
        ]
      }
    },
    "/api/check-password/{dn}": {
      "post": {
        "description": "Verify a password",
        "operationId": "post_check_password",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Check",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Post Check Password",
                  "type": "boolean"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Check Password",
        "tags": [
          "Editing"
        ]
      }
    },
    "/api/entry/{dn}": {
      "delete": {
        "operationId": "delete_entry",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Delete Entry",
        "tags": [
          "Editing"
        ]
      },
      "get": {
        "description": "Retrieve a directory entry by DN",
        "operationId": "get_entry",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Entry"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Get Entry",
        "tags": [
          "Editing"
        ]
      },
      "post": {
        "operationId": "post_entry",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "title": "Attributes",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangedAttributes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Post Entry",
        "tags": [
          "Editing"
        ]
      },
      "put": {
        "operationId": "put_entry",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "title": "Attributes",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangedAttributes"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Put Entry",
        "tags": [
          "Editing"
        ]
      }
    },
    "/api/ldif": {
      "post": {
        "description": "Import LDIF",
        "operationId": "post_ldif",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Ldif",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Upload Ldif",
        "tags": [
          "Editing"
        ]
      }
    },
    "/api/range/{attribute}": {
      "get": {
        "description": "List all values for a numeric attribute of an objectClass like uidNumber or gidNumber",
        "operationId": "get_range",
        "parameters": [
          {
            "in": "path",
            "name": "attribute",
            "required": true,
            "schema": {
              "title": "Attribute",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Range"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Attribute Range",
        "tags": [
          "Misc"
        ]
      }
    },
    "/api/rename/{dn}": {
      "post": {
        "description": "Rename an entry",
        "operationId": "post_rename_entry",
        "parameters": [
          {
            "in": "path",
            "name": "dn",
            "required": true,
            "schema": {
              "title": "Dn",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": "Rdn",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Rename Entry",
        "tags": [
          "Editing"
        ]
      }
    },
    "/api/schema": {
      "get": {
        "description": "Dump the LDAP schema as JSON",
        "operationId": "get_schema",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schema"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Ldap Schema",
        "tags": [
          "Misc"
        ]
      }
    },
    "/api/search/{query}": {
      "get": {
        "description": "Search the directory",
        "operationId": "search",
        "parameters": [
          {
            "in": "path",
            "name": "query",
            "required": true,
            "schema": {
              "title": "Query",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Search",
        "tags": [
          "Navigation"
        ]
      }
    },
...
    "/api/whoami": {
      "get": {
        "description": "DN of the current user",
        "operationId": "get_who_am_i",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Who Am I",
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBasic": []
          }
        ],
        "summary": "Whoami",
        "tags": [
          "Misc"
        ]
      }
    }
  }
}